Unify Tequilapi server & client DTOs for location
